requirements_frozen.txt
ERROR: Error [Errno 2] No such file or directory: 'git' while executing command git config --get-regexp 'remote\..*\.url'
WARNING: cannot determine version of editable source in /oscar/home/hnam16/hazel/code/cjepa/stable-pretraining (git command not found in path)
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.4.0
av==16.0.1
cattrs==25.3.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.2
cycler==0.12.1
datasets==4.0.0
dill==0.3.8
einops==0.8.1
exceptiongroup==1.3.0
Farama-Notifications==0.0.4
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
gymnasium==1.2.2
h5py==3.15.1
hf-xet==1.2.0
huggingface-hub==0.36.0
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lightning==2.5.6
lightning-utilities==0.15.2
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
minari==0.5.3
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.1
pyarrow==20.0.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.6
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-cache==1.2.1
rich==14.2.0
richuru==0.1.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentry-sdk==2.45.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
-e /oscar/home/hnam16/hazel/code/cjepa/stable-pretraining
submitit==1.5.3
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
timm==1.0.22
tokenizers==0.22.1
torch==2.9.1
torchcodec==0.8.1
torchmetrics==1.8.2
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.1
triton==3.5.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.5.0
wandb==0.23.0
wcwidth==0.2.14
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
